v2/code/requirements.txt
requests~=2.28.1
PyYAML~=6.0
packaging~=21.3
pandas~=1.5.2
seaborn~=0.12.1
matplotlib~=3.6.2
python-dateutil~=2.8.2
johnnydep~=1.17.5
pip~=22.3.1
pytest~=7.2.1
numpy~=1.24.1
scipy~=1.9.3
dataclasses~=0.6
munch~=2.5.0
nltk~=3.8.1
scikit-learn~=1.2.2